Splits up alpha page tests per component
